requirements.txt
streamlit==1.40.0
supabase==1.0.4
python-dotenv==1.0.0
pytz==2023.3
pandas==2.1.2
streamlit-calendar==0.5.0